Added a few features
"Space-aware": If there is 1+ empty lines above/below a region, unmarking will result in one empty line above/below, instead of the two it left before.

"Quick-unmark": You can now "quickly" unmark if your cursor is within a region, whereas before the entire region needed to be selected.

"showInRuler": Regions can now be show to the right of the editor/minimap in the "ruler", or scrollbar.

Also:

Changed configuration image in `README.md` to text so it can be selected.